Renamed contants files to be post fixed with -enum. Removed the constant style naming from the includes of the constant arrays. Moved the reference number regex into the reference-number-enum.
