+++ app/constants/reference-number-enum.js
@@ -1,0 +1,3 @@
+exports.VALID_CHARACTERS = '0123456789ABCDEFGHJKMNPQRSTVWXYZ'
+exports.VALID_LENGTH = 7
+exports.IS_VALID_REGEX = new RegExp('[' + this.VALID_CHARACTERS + ']{7}')
