--- app/constants/reference-number.js
@@ -1,4 +1,0 @@
-module.exports = {
-  REFERENCE_NUMBER_CHARACTERS: '0123456789ABCDEFGHJKMNPQRSTVWXYZ',
-  REFERENCE_NUMBER_LENGTH: 7
-}
